Add additional columns to upload output, move config to function inputs
The big change here is that the data_grouper_freq, segment_columns, and
whylabs_dataset_id secrets are no longer needed. That config is now just
duplicated for each row of the input. It's a lot easier to specify these
as inputs (at the cost of memory) than it is to have to create multiple
UDF/integrations for each time that uses a different dataset id.